+++ packages/components/src/image-preview/index.ts
@@ -1,0 +1,1 @@
+export { ImagePreview, type ImagePreviewProps } from './image-preview'
